Update Babel configuration for Node 6
Node 6 is the officially supported minimum version for this package.
We can update the browserlist to this Node version to make sure
that the build output is compatible with Node 6.

Since this package does not use any APIs from Node 6+, the build
output becomes significantly smaller. Even more so, it makes the
Babel runtime unnecessary, since all of the APIs already run on
Node 6.

Therefore, we can also remove the dependency on the Babel runtime,
which saves ~7MB for downstream users.

Then, some of the tests had to be updated to use the correct
Jest 24 APIs. It's not clear to me why they weren't failing before,
but when I ran `npm i` and ran Jest, it started complaining
about incorrect matchers. For example, `toBeTruthy()` does not
take any argument per the Jest docs:
https://archive.jestjs.io/docs/en/24.x/expect#tobetruthy

Additionally, the `.toBeA` function does not exist in Jest 24.
Instead, `.toBeInstanceOf` is the relevant matcher:
https://archive.jestjs.io/docs/en/24.x/expect#tobeinstanceofclass

Lastly, the `.toExcludeKey` does not exist and instead, Jest
has the `.not.toHaveProperty` matcher:
https://archive.jestjs.io/docs/en/24.x/expect#tohavepropertykeypath-value

Fixes #158