Update Babel configuration for Node 6
Node 6 is the officially supported minimum version for this package.
We can update the browserlist to this Node version to make sure
that the build output is compatible with Node 6.

Since this package does not use any APIs from Node 6+, the build
output becomes significantly smaller. Even more so, it makes the
Babel runtime unnecessary, since all of the APIs already run on
Node 6.

Therefore, we can also remove the dependency on the Babel runtime,
which saves ~7MB for downstream users.

Fixes #158